requirements.txt
# Web Framework
flask>=3.0.0,<4.0.0
flask-login>=0.6.0,<1.0.0
flask-sqlalchemy>=3.1.0,<4.0.0
werkzeug>=3.0.0,<4.0.0
flask-mail>=0.10.0,<1.0.0
gunicorn>=21.0.0,<23.0.0

# AI/ML Dependencies
openai>=1.0.0,<2.0.0
google-generativeai>=0.8.0,<1.0.0
sentence-transformers>=2.2.0,<3.0.0
langchain>=0.1.0,<1.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
faiss-cpu>=1.7.0,<2.0.0

# Document Processing
pymupdf>=1.23.0,<2.0.0

# Web Automation (only if needed)
selenium>=4.16.0,<5.0.0
webdriver-manager>=4.0.0,<5.0.0

# System Monitoring
schedule>=1.2.0,<2.0.0
psutil>=5.9.0,<6.0.0

# PDF Generation
WeasyPrint>=59.0

reportlab>=3.6.12

Flask-Migrate==4.0.4

PyPDF2

psycopg2-binary>=2.9.0,<3.0.0

# For data backup
openpyxl

requests==2.31.0

beautifulsoup4
lxml

tavily-python

google-genai